fix(torch-frontend): Avoids importing backend helper to frontend
