+++ ivy/functional/backends/torch/layers.py
@@ -8,7 +8,7 @@ import torch
 import ivy
 from ivy.func_wrapper import with_unsupported_dtypes, with_supported_dtypes
 from . import backend_version
-from ivy.functional.ivy.layers import _handle_padding, _deconv_length
+from ivy.functional.ivy.layers import _get_embed_dim, _handle_padding, _deconv_length
 
 
 @with_supported_dtypes(
@@ -114,19 +114,6 @@ multi_head_attention.partial_mixed_handler = (
     )
     == 1
 )
-
-
-def _get_embed_dim(
-    in_proj_weights, q_proj_weights, k_proj_weights, v_proj_weights, query
-):
-    pre_embed_dim = query.shape[-1]
-    if ivy.exists(in_proj_weights):
-        embed_dim = in_proj_weights.shape[0] / 3
-    elif all([ivy.exists(x) for x in [q_proj_weights, k_proj_weights, v_proj_weights]]):
-        embed_dim = q_proj_weights.shape[0]
-    else:
-        embed_dim = None
-    return pre_embed_dim, embed_dim
 
 
 @with_unsupported_dtypes(
