+++ ivy_tests/test_ivy/test_frontends/test_torch/test_nn/test_functional/test_layer_functions.py
@@ -3,7 +3,7 @@ from hypothesis import strategies as st
 
 # local
 import ivy
-from ivy.functional.backends.torch.layers import _get_embed_dim
+from ivy.functional.ivy.layers import _get_embed_dim
 from ivy_tests.test_ivy import helpers
 from ivy_tests.test_ivy.helpers import handle_frontend_test
 from ivy_tests.test_ivy.test_functional.test_nn.test_layers import _mha_helper
